requirements.txt
pandas>=2.0
faker>=40.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0